pyproject.toml
[project]
name = "base-research-repo"
version = "0.1.0"
description = "Base research repository for BAILab - A comprehensive framework for ML research"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.8.1",
    "datasets>=2.12.0",
    "matplotlib>=3.10.3",
    "nltk>=3.9.1",
    "numpy>=2.3.1",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
    "scikit-learn>=1.7.0",
    "seaborn>=0.13.2",
    "tokenizers>=0.13.0",
    "torch>=2.7.1",
    "tqdm>=4.65.0",
    "transformers>=4.53.2",
    "wandb>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
